PubSub_Streaming_Project/requirements.txt
google-cloud-pubsub==2.7.1
google-cloud-storage==1.42.3